firebase realtime db sync Feed

+++ Udemy/Dorisgram/app/src/main/java/com/kimmandoo/Feed.kt
@@ -4,7 +4,7 @@ data class Feed(
     val userId: String,
     val imageUrl: String,
     val profileImageUrl: String,
-    var likeCount: Int,
+    var likeCount: Long,
     var isLike: Boolean,
     var isBookmark: Boolean
 )
